Added files to new dir
